fix: apply review changes
