+++ static/preview.html
@@ -55,6 +55,6 @@
       })
     </script>
   </body>
-  <script src="@/unity/Build/UnityLoader.js"></script>
-  <script defer async src="@/artifacts/preview.js"></script>
+  <script src="/@/unity/Build/UnityLoader.js"></script>
+  <script defer async src="/@/artifacts/preview.js"></script>
 </html>
